+++ docs-src/components/utils/Utils.svelte
@@ -22,7 +22,7 @@
 	<p>
 		<em>*</em> <a href="https://svelte.dev/docs/svelte-components#script-4-prefix-stores-with-$-to-access-their-values">
 		svelte store variables</a> - when reading the value, add <em>$</em> to the name, e.g.
-		<em>$FOCUSABLE_SELECTOR</em>.
+		<em>$ANIMATION_SPEED</em>.
 	</p>
 </div>