+++ docs-src/components/utils/properties/focusable-selector.svelte
@@ -1,6 +1,5 @@
 <Util id="FocusableSelector" name="FOCUSABLE_SELECTOR" {example}>
 	<ul>
-		<li>Svelte store<em>*</em>
 		<li>Type: string
 		<li>Returns a list of selectors that can be focused.
 	</ul>
@@ -13,7 +12,7 @@ const example = `
 <script>
 	import { FOCUSABLE_SELECTOR } from '@perfectthings/ui';
 
-	const focusableElements = document.querySelectorAll($FOCUSABLE_SELECTOR);
+	const focusableElements = document.querySelectorAll(FOCUSABLE_SELECTOR);
 	console.log(focusableElements);
 &lt;/script>
 `;