+++ docs-src/components/utils/properties/index.js
@@ -1,2 +1,3 @@
+export { default as AnimationSpeed } from './animation-speed.svelte';
 export { default as FocusableSelector } from './focusable-selector.svelte';
 export { default as PrefersDark } from './prefers-dark.svelte';